agentic_codebase_genius/backend/requirements.txt
jaseci
jaclang
gitpython
networkx
matplotlib
